Change health int to float in level spawner

+++ Assets/Source/Scripts/Targets/Target.cs
@@ -67,7 +67,7 @@ public class Target : MonoBehaviour, IPoolObject<Target>
         transform.rotation = rotation;
     }
 
-    public void ResetSettings(int health)
+    public void ResetSettings(float health)
     {
         _healthView.SetMaxHealth(health);
         _healthView.Restore();
